requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# OpenAI y LangChain
openai>=1.3.5
langchain>=0.1.20
langchain-openai>=0.1.8
langchain-community>=0.0.38

# Configuración
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Base de datos
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.10
pgvector>=0.2.3

# Ciencia de datos (versiones compatibles con Python 3.13)
numpy>=1.26.0
scikit-learn>=1.3.2
pandas>=2.1.3

# Utilidades
python-multipart>=0.0.6
aiofiles>=23.2.1
setuptools>=68.0.0

